% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcana-class.R
\name{arcana}
\alias{arcana}
\title{The Arcana Class}
\usage{
arcana(
  name = character(0),
  number = integer(0),
  emoji = character(0),
  keywords = list()
)
}
\arguments{
\item{name}{The name of the card (e.g., "The Fool").}

\item{number}{The cards number (an integer), from 0 to 21.}

\item{emoji}{The emoji associated with the card.}

\item{keywords}{A comma-separated character string of keywords. These will
be automatically converted to a character vector.}
}
\value{
An \code{S7} object of class \code{arcana}.
}
\description{
The \code{arcana} class defines a tarot card. It uses the \code{S7} object-oriented
framework to ensure that each card object has a consistent structure with
validated properties.
}
\examples{
 card <- major_arcana_list$the_fool
 the_fool <- arcana(
   name = card$name,
   number = card$number,
   emoji = card$emoji,
   keywords = list(
     upright = card$upright,
     reversed = card$reversed
   )
 )

}
