% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{major_arcana_data}
\alias{major_arcana_data}
\title{Major Arcana Tarot Card Data}
\format{
A list with 22 items and 4 items each:
\describe{
\item{number}{Card number (0-21)}
\item{name}{Card name}
\item{upright_keywords}{Keywords for upright orientation}
\item{reversed_keywords}{Keywords for reversed orientation}
}
}
\usage{
major_arcana_data
}
\description{
A list object containing the 22 Major Arcana tarot cards with their
upright and reversed keyword meanings.
}
\keyword{datasets}
