% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{new_arcana}
\alias{new_arcana}
\title{Create an Arcana Object}
\usage{
new_arcana(name, orientation, keywords, number)
}
\arguments{
\item{name}{Character string, name of the tarot card}

\item{orientation}{Character string, either "upright" or "reversed"}

\item{keywords}{Character string, comma-separated keywords for the card}

\item{number}{Integer, the card number (0-21)}
}
\value{
An object of class "arcana"
}
\description{
Constructor function for the arcana S3 class.
}
